Refactor Android notifications
